refactor(contractor-onboarding) - extract contract details schema hook to common/api
Moves useContractorOnboardingDetailsSchemaWithCurrencies (and its private
helpers) out of ContractorOnboarding/api.ts into a shared
src/common/api/contractor-contract-details.ts as
useContractorContractDetailsSchema, so it can be reused by other flows
that need the same contractor contract details schema.
